CS: Property meta - Property model - Property parameter container - added field immutable in model 'Field' - common parameter validator - property parameter validator

+++ config-provider/src/main/java/com/kostylenko/config_service/config_provider/common_config/Properties.java
@@ -5,7 +5,7 @@ import java.util.Map;
 
 public interface Properties {
 
-    Integer getInteger(String name);
+    Float getFloat(String name);
 
     Boolean getBoolean(String name);
